+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.2] - 2019-08-12
+### Changed
+- Renamed command method to execute
+- Renamed run method to execute
+
 ## [1.0.1] - 2019-08-07
 ### Changed
 - Pass args to command in call for Simple based commands